[NetlistWriter] Fixed Tcq SDF Writing for BlackBoxs
The Netlist Writer is used to write the post-implementation netlist and
SDF files for use in post-implementation timing analysis. While working
with this tool on blocks with multi-port, registered output pins, I
found that the SDF file generated was not correct.

The issue was that the Netlist Writer always assumed that any registered
output ports had only 1 bit, so the SDF file did not contain delays per
output port causing it to be invalid.

This was easy to fix by making clock to q paths timing arcs instead so
that the pin IDs in each port can be passed to the SDF writer and the
SDF writer can annotate the correct delays.