[AP] Generalized Argument Parsing and Added Target Density
The `appack_max_dist_th` cli option has a useful regex based interface
which can be used in other parts of VPR, especially in AP. Generalized
the interface so it can be used outside of the context of max distance
thresholding.

Used this general interface to pass in a target density per physical
tile type. This spoofs the capacity into being smaller than it appears
which tells the partial legalizer to not fill the tiles beyond the
capacity.

Setting the target density to interesting values revealed a bug in the
partial legalizer regarding numerical precision when computing the
prefix sum. Changed the underlying data type within the prefix sum class
used into fixed-point which resolved the issue.