[AP] Marking Global Nets in AP
After investigating how the placer and packer decide which nets are
global, I realized that we were not marking any nets as global. The Atom
Netlist annotates nets as global after packing, when loading the
clustered netlist. It decides if a net is global or not if the net
enters a global port in the clustered netlist.

We can replicate this behaviour during AP by looking at the Atom Netlist
and doing something similar. This allows us to tell if a net is global
or not.

In AP, we always want to ignore global nets since their cost does not
play a role on the overall objective. By ignoring these nets, we can
further improve the performance and quality of the solution.

Now that we know which nets are global, I have also updated the
wirelength estimator in the AP flow to make it align better with the
placer's wirelength estimator. They now line up very closely, but are
still slightly off due to some factors being unable to be ported over
easily.