[Arch] Added Zero ASIC's Z1000 eFPGA Architecture
Zero ASIC uses VPR as part of its open-source FPGA CAD flow (Logik).
Because of this, its architecture XML file is also open-source.

This brings in Zero ASIC's Z1000 architecture. A 2048 4-LUT/FF
architecture with 1024 GPIOs and can support up to 4 unique clock
domains.

I had to slightly modify the architecture file since the Z1000 uses
custom register primitives which its synthesis front-end can target
instead of regular ".latch" primitives. To get around this, I added a
custom mode to the BLE to use the ".latch" primitive instead of the
register primitives. Since the pins for the ".latch" primitive and the
register primitives have one-to-one correspondance, this should have no
affect on VPR's execution.

In order for these circuits to route properly, the z1000 RR graph must
also be provided. Included these as zip files and added a make command
which will extract the RR graphs. Zipped, this file is around 2 MB,
unzipped it is 90 MB.

Added a tasks to NightlyTest7 which runs the MCNC benchmarks that could
fit on this device. There is currently an issue in the packer causing it
to give up too early and not pack as much as it can; the circuits
affected by this bug have been commented out and will be uncommented
once that bug is fixed.