[AP][DP] Added Option to Scale Init T in Annealer
Since the AP flow produces an initial placement that has 4x improved
estimated wirelength and 2x improved CPD, the initial temperature of the
annealer is too high. From experimentation, I have found that reducing
the initial temperature in half improves quality slightly and reduces
runtime.

Added an option to the Placer to scale up or down the initial
temperature. The default flow just sets this scale to 1.0, but the AP
flow sets this scale to 0.5 currently.

I have also exposed this option to the command-line so this option can
be swept more easily.