[Packer] Reorganized Iterative Packer Algorithm
The packer iteratively restarts itself with different options if the
packer fails to find a dense enough clustering. The way this was written
was very hard to follow and was error prone.

Rewrote the iterative packer code to be more like a state-machine. This
makes it easier to add new states in the future and maintain the current
states.