[Packer] Fixed NightlyTest Failure with Tight Floorplan Constr.
The very tight floorplan constraints were failing after the recent
changes to the iterative packing algorithm. After investigation I found
that there is a bug with unrelated clustering and RAMs which is causing
floorplan constraints to not be adhered to. The original code ignored
this issue by not doing unrelated clustering when there are floorplan
constraints.

To fix this issue, reimplemented what the code was originally doing to
allow this testcase to proceed. This just moves the unrelated clustering
to after trying all floorplan region options.