[Build] Updated VTR Versioning
The way VTR was setting its version information was slightly
non-standard and was potentially causing issues when VTR was being built
using sudo and then used by a non-sudoer. The reason this process is so
complicated is because we want the versioning information to change if
anything is modified in the repository or a new commit is created.

Ryan Pavlik provides a CMake module to do exactly this, which is also
used by the OpenROAD project. I updated our build process to use this
CMake module.

This also resolved an issue where VTR's version was printing v9.0.0;
while our revision was showing v8.0.0. Now they both show v9.0.0.