[CI] Moved Sanitized Test from Debug to RelWithDebInfo
I expect the sanitized test to be the tall pole for the speed of the CI
due to its long run time on the testcases. After investigating with
Vaughn, we realized that this test was using the debug build of VTR.
This was likely making it much slower than it needed to be.

To improve the speed of this test, moved it to use the release build. We
still want symbols to give good error messages.