[AP][MassLegalizer] Resolved Issues in Partitioner
Found some low-hanging fruit in the partial legalizer regarding how we
merge the windows and how we partition the blocks.

I found that the windows were not being fully merged (i.e. there still
existed overlap between windows). I updated the algorithm to iteratively
retry to merge if the bounding box of the window changes. To maintain
good performance, I added a spatial lookup to narrow down the region
where overlap can occur.

The partitioning of blocks had a major bug in it where it was not
actually moving blocks over most of the time. This was causing the
partial legalizer to struggle to resolve of a legal solution. To resolve
this, I completely rewrote the partitioning algorithm to make it more
stable. The new algorithm starts by putting the blocks on the side they
want (if their solved position is on the left of the partition, the will
go into the left partition). It then moves the blocks closest to the
partition line until the overfill of both partitions is as balanced as
can be.

These two optimizations together seem to have improved WL by around 2%
and CPD by around 1%.