[Build] Locked Version Tag Reference
To get a unique version for every commit on Master, git describe is
used. This produces a unique (monotonically increasing) string by
getting the number of commits since a tag.

We originally let git decide which tag to use, which I assumed was the
most recent tag; however, some users have been noticing that the version
has been relative to different tags (like v9.0.0-candidate). In the
future, we may want to also control when we want to move the tag this
version is relative to manually, instead of letting Git decide.

To make this more robust and controllable, forced the git describe to
be relative to a specific tag which is set in VTR's CMake (similar to
the major version number).