[Place] Added an Initial Placement Checkpoint
I found that the annealer may sometimes pick a temperature so high that
it destroys the quality of the initial placement. This behavior is most
aparent with the AP flow, since its initial placement quality is much
better than the default flow.

Added a checkpoint at the start of the anneal that saves the initial
placement and, after 20 iterations, if the quality of the current anneal
is worse than the initial placement, it will restore that placement and
continue at that temperature.