[APPack] Tuned Up APPack Code
This is the second PR that brings in private tunings for the Triple-AP
work into VTR master.

These are all the tuning changes to the APPack algorithm. This comes
with changes to the max displacement threshold values, improved
fall-back logic, and better unrelated clustering logic.

Also fixed a small issue in the delay derivative calculation of the
analytical solver to explicitly handle boundary conditions of the
derivative.