[APPack] Initial Placement Fallback
This is the first PR bringing back changes from the AP tuning branch
into the main branch of VTR.

This PR adds the ability for the AP initial placer of APPack to fallback
on the original initial placer implementation if it fails to find a
valid initial placement. This only occured in two testcases from the
Koios benchmark suite.