[3D] Added Basic Test and Fixed Bugs in 3D Flow
While I was experimenting with the 3D flow, I ran into some simple
issues in the placement code for a basic 3D FPGA. The issue is that this
FPGA has all of the IO tiles on one die and all the CLBs on another. The
code was not handling the case when no tile exists on a given die and
was hitting assertions and seg faults.

I have fixed these issues and added a basic test to exercise this on a
set of basic circuits.