[AP][Solver] Added The Identity Analytical Solver
While testing the AP flow, it is convenient to only focus on one stage
of the flow at a time. It would be nice to have a very simplified
version of the stages which we know will always work.

I have created a new solver called the "identity" solver which acts as
an identity matrix where it just passes the solution from the prior
partial legalizer iteration along. This allows me to focus on the
partial legalizer without having to worry about the solver as I make
more substantial changes while working on new applications of the AP
flow.