[AP][PartialLegalizer] Added None Partial Legalizer
When testing the AP flow, it can be convenient to have a partial
legalizer which does literally nothing. This allows us to debug other
parts of the flow without worrying about the partial legalizer.

I have added the "none" partial legalizer which acts as a placeholder
and does literally nothing.