[LibVTR][Math] Fixed Issues with Median Calculation
The median calculation in vtr::math was not working for the double
precision type. I think it was working for the floating point type due
to how the code was written, but I am not sure.

The issue with the median calculation is whether last is included in the
set or not when computing the median. Usually it is not, so I updated
the code accordingly.

I also cleaned up the code to make it a bit more user friendly.

To lock in these changes I have added unit tests.