[Logging][Warnings] Suppressed Overly Verbose Warnings
There were many occurences of warning messages that were either overly
verbose or unecessary.

In the RR graph reading code, there were several warnings that were
generated when nodes could not be found for a cost index. This warning
is good to keep around I think; however we do not need to print the
warning for each cost index.

In check netlist, we were printing a warning when the netlist does not
contain any global nets to non-global pins, which is a strange thing to
warn about.

In the annealer, we have been warning whenever the initial temperature
was being calculated. This warning was saying that some of the swaps
attempted was rejected; but I do not think this needs to be warned
about.

In the router, there were two warnings which are not of any use to a
user, but for developers they may be useful. I have raised the verbosity
level of these warnings. This required that I update the router code to
pass around the route_verbosity variable.