Fix regression testing false negative
The QoR parsing script returned -1 in case of a mismatch between arch file name in
the actual run and the golden results. If there was 1 run failures, the -1 would
cancel the 1 and as a result the regression tests would pass while having errors.
This commit fixes that by not throwing an exception in that specific case (Therefore
not returning -1) and taking the absolute value of the QoR and run failures.