[Place] Estimating Starting T Using Equilibrium T
I found that the way that the annealer estimates the initial temperature
to be too high when the initial placement is of very good quality (for
example, after AP).

Added a new way of estimating the starting temperature by
setting it to an estimation of the equilibrium temperature. The
equilibrium temperature is the temperature at which the change in cost
after an annealing iteration would be 0.

The old way (of using the variance of the change in cost) is still the
default; however, this new method can be turned on in the command-line.