[Pack][Legalizer] Added get_cluster_name Method
I am trying to reduce the exposed state from the cluster legalizer, in
particular how t_pb pointers are being used outside of the legalizer
class.

I found a very easy cleanup where a few of the usages were just getting
the name of the cluster. This makes sense to expose through an API.