[Pack] Removed Temp Scratch Pad from pb_graph_node
Each pb_graph_node kept a void* called temp_scratch_pad. This was used
to store temporary data during packing. This is incredibly dangerous and
also very wasteful on space.

I found that this was only used by the prepacker, so it was easily
replaced by an unordered map which accomplished the same thing.