[CI] Fix CPP Format Issue
The format test for CPP was done through a bash script which calls make
format under the hood. The issue was that if make format fails, it
thinks that the formatting is good, which is bad. I have fixed the
script so that it actually fails if make format fails.

The thing that was causing make format to fail was that the dependencies
that VPR needs to build were not installed. Added these dependencies.