[Pack] Logical RAM Mapping: Adding timing pass to the end of area pass.
It is doing an initial assignment to each logical RAM (the smallest type).

Then based on the area cost, we remap them to the smaller area consuming type
(proxy for area of a physical block is its height*widht).

Then for the most timing critical logical rams, we try to remap them to the smaller
type. Smaller types believed to be faster but for generalizability this should be
changed to use the actual internal delay of the RAM. Also, this timing remapping should
be skipped for the RAMs whose output is registered (this skip is not in the code yet).