[Spelling] Fixed Simple Spelling Mistakes in Route Dir
This is the last of the simple spelling mistakes in the VPR directory.
The rest are variable name changes which are not hard, but should be
isolated into their own commits.

NOTE: The rr_graph.cpp and rr_graph2.cpp files were left behind to
prevent issues with currently open PRs.