[CI] Added Spell Checking to CI
Cleaned up the last necessary spelling mistakes in VPR so we can now add
codespell to the CI. This will add spell-checking to the CI tests so
when people merge code into VTR, we can ensure that there are no obvious
spelling mistakes.

I have added a config file which contains which files should be ignored
and what words can be ignored.