[Spelling] Fixed Spelling Mistakes in VTR Flow
Fixed the spelling mistakes in the VTR flow. We exclude the pearl
libraries and benchtracker since these projects are not being maintained
at this time.

We also ignore the benchmarks and tasks directories since these are
mainly used for testing and contain a lot of false positives. We do not
want to make it challenging to add more tests.