[Benchmarks] Updated S10 Other Benchmarks
A few S10 benchmarks were included directly into VTR for use in the
strong tests.

The S10 benchmarks were recently updated, but these tests were not
updated in the repo. This causes a conflict when the Titan benchmarks
are downloaded, where these benchmarks are marked as dirty.

Pushing the most recent version of these benchmarks.